requirements.txt
google-ads>=23.0.0
